+++ Assets/Scripts/Glaidiator/Model/Actions/Attack.cs
@@ -9,7 +9,7 @@
         // returns false if timer is missing/done/not started yet
         // returns true if timer is ticking
         // TODO: add hitbox info
-        public Attack(int id, string name, float damage, int cost, bool canMove, bool canAction, float actionDuration, float cooldownDuration = 0f) 
+        public Attack(int id, string name, float damage, float cost, bool canMove, bool canAction, float actionDuration, float cooldownDuration = 0f) 
             : base(id, cost, canMove, canAction, actionDuration)
         {
             Damage = damage;
